+++ lib/src/error.dart
@@ -1,18 +1,39 @@
+import 'dart:ui';
+
+import 'package:firebase_crashlytics/firebase_crashlytics.dart';
 import 'package:flutter/material.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 import 'package:sdui/sdui.dart' as sdui;
 import 'package:http/http.dart';
-import 'package:wutsi_wallet/src/environment.dart';
 import 'package:wutsi_wallet/src/login.dart';
 import 'package:wutsi_wallet/src/access_token.dart';
 
 import 'device.dart';
 
-Environment environment = Environment(Environment.defaultEnvironment);
-
-void initError() async {
-  environment = await Environment.get();
+void setupErrorHandling() async {
   sdui.sduiErrorState = (context, error) => SDUIErrorWidget(error: error).build(context);
+
+  // Flutter Screen of Death
+  ErrorWidget.builder = (FlutterErrorDetails details) {
+    return FlutterErrorWidget(details: details);
+  };
+  FlutterError.onError = (details){
+    // Default behaviour
+    FlutterError.presentError(details);
+
+    // Push errors Crashlytics
+    if (FirebaseCrashlytics.instance.isCrashlyticsCollectionEnabled){
+      FirebaseCrashlytics.instance.recordFlutterError(details, fatal: true);
+    }
+  };
+
+  PlatformDispatcher.instance.onError = (error, stack) {
+    if (FirebaseCrashlytics.instance.isCrashlyticsCollectionEnabled){
+      FirebaseCrashlytics.instance.recordError(error, stack, fatal: true);
+    }
+    return true;
+  };
+
 }
 
 ///
@@ -64,7 +85,6 @@ class SDUIErrorWidget extends StatelessWidget{
             if (snapshot.hasData){
               return sdui.DynamicRoute(
                 provider: LoginContentProvider(context,
-                    environment,
                     phoneNumber: snapshot.data?.phoneNumber(),
                     hideBackButton: true
                 ),