+++ lib/src/login.dart
@@ -5,15 +5,16 @@ import 'package:wutsi_wallet/src/environment.dart';
 /// Login Page
 class LoginContentProvider implements RouteContentProvider {
   final BuildContext context;
-  final Environment environment;
   String? phoneNumber;
   bool? hideBackButton;
 
-  LoginContentProvider(this.context, this.environment, {this.phoneNumber, this.hideBackButton});
+  LoginContentProvider(this.context, {this.phoneNumber, this.hideBackButton});
 
   @override
-  Future<String> getContent() => Http.getInstance()
-      .post(loginUrl(phoneNumber, hideBackButton ?? false, environment), null);
+  Future<String> getContent() => Environment.get()
+      .then((value) => Http.getInstance()
+        .post(loginUrl(phoneNumber, hideBackButton ?? false, value), null)
+  );
 
 
   static String loginUrl(String? phoneNumber, bool hideBackButton, Environment environment) =>